fix(common): split executor into submodule
The executor and executor_factory flat files were organized such that
executor_factory imported BaseBlockExecutor via the crate root re-export
rather than directly from its sibling module. Splitting into an executor/
submodule with dedicated result, factory, and executor files resolves
this implicit dependency through the public API surface and aligns with
the module organization conventions in CLAUDE.md.
